before VPS

+++ wafweb/wafwebapp/classes/loader/neoloader.py
@@ -7,6 +7,7 @@ from geotoneo import GeoToNeo
 import time
 from waflog import WafLog
 from spatialapi import SpatialApi
+from classes.neo_classes.province import Province
   
 
 
@@ -18,19 +19,24 @@ class NeoLoader:
         self.neodb = GeoToNeo()
         self.db = DbConnection()
         
-    def initContinent(self, language):
+    def initContinent(self, language ='en'):
         table_continent = "dn_continent_en_neo"
         fields_continent = "id, lower(name), lower(short_name), lower(language)"
         where_continent = """language = '{0}'  """.format(language)
         self.continent = self.db.execute(fields=fields_continent, table=table_continent, where=where_continent, one=False)
     
-    def initRegion(self, language):
+    def initRegion(self, language ='en'):
         table_continent = "dn_region_en_neo"
         fields_continent = "id, lower(name), lower(short_name), lower(country_list), lower(language)"
         where_continent = """language = '{0}' and root = 'True'  """.format(language)
         self.region = self.db.execute(fields=fields_continent, table=table_continent, where=where_continent, one=False)
         
-    def initCountry(self, language):
+    def initProvince(self, language ='en'):
+        table_continent = "dn_province_en_neo"
+        fields_continent = "id, lower(name), lower(country_code)"
+        self.province = self.db.execute(fields=fields_continent, table=table_continent, where=None, one=False)
+        
+    def initCountry(self, language ='en'):
         if language  =='en':
             table_country = "dn_country_en_neo"
             fields_contry = "id, lower(country_code),lower(name), lower(continent)"
@@ -42,7 +48,7 @@ class NeoLoader:
         
     def initCity(self):
         table_city = "dn_city_en_neo"
-        fields_city = "id, lower(name) ,lower(country_code), lat, lan"
+        fields_city = "id, lower(name) ,lower(country_code), lat, lan, lower(province)"
         self.city_root = self.db.execute(fields=fields_city, table=table_city, one=False)
         
     def initCityHu(self):
@@ -79,7 +85,6 @@ class NeoLoader:
                     WafLog().neologger.info(i)
                     
     def loadRegion(self):
-        self.initCity()
         WafLog().neologger.info('start to create regions')
         i = 0
         start_time = time.time()
@@ -88,6 +93,21 @@ class NeoLoader:
         for region in self.region:
             reg = Region(region[0],region[1], region[2], region[3],region[4])
             self.neodb.createRegion(reg)
+            i+=1
+            if i%500 ==0:
+                elapsed_time = time.time() - start_time
+                WafLog().neologger.info("n: {0} elapsed: {1}".format(i,elapsed_time))
+                start_time = time.time()
+    def loadProvince(self):
+        self.initProvince()
+        WafLog().neologger.info('start to create provinces')
+        i = 0
+        start_time = time.time()
+        # your code
+        
+        for province in self.province:
+            provi = Province(province[0],province[1], province[2])
+            self.neodb.createProvince(provi)
             i+=1
             if i%500 ==0:
                 elapsed_time = time.time() - start_time
@@ -102,7 +122,7 @@ class NeoLoader:
         # your code
         
         for city in self.city_root:
-            cit = City(city[0],city[1], city[2], city[3], city[4])
+            cit = City(city[0],city[1], city[2], city[3], city[4], city[5])
             self.neodb.createCity(cit)
             i+=1
             if i%500 ==0:
@@ -150,18 +170,27 @@ class NeoLoader:
         self.neodb.createIndexOnCity()
     def createIndexOnGeo(self):
         self.neodb.createIndexOnGeo()
+    def createIndexOnProvince(self):
+        self.neodb.createIndexOnProvince()
            
 if __name__ == "__main__":
     continents = {'en'}
     countries = {'en'}
+    drop = True
     neotest = NeoLoader()
-    spatialapi = SpatialApi() 
-    neotest.deleteDB()
+    spatialapi = SpatialApi()
+    if drop: 
+        neotest.deleteDB()
     neotest.loadContinent(continents)
+    
     neotest.loadCountry(countries)
     neotest.createIndexOnCountry()
     neotest.connectCountries()
+    
     neotest.loadRegion()
+    neotest.loadProvince()
+    neotest.createIndexOnProvince()
+    
     neotest.loadCity()    
     neotest.createIndexOnCity()  
     neotest.createIndexOnGeo()
